feat: add rdbms storage (#56)
* feat: add rdbms storage

Closes #1


* fix: make created on property read and write

+++ src/Http/AbstractRedirectMiddleware.php
@@ -19,7 +19,7 @@ abstract readonly class AbstractRedirectMiddleware implements MiddlewareInterfac
     {
         return $this->responseFactory
             ->createResponse(
-                HttpResponseCodeEnum::TemporaryRedirect->value,
+                (int) HttpResponseCodeEnum::TemporaryRedirect->value,
                 HttpResponseCodeEnum::TemporaryRedirect->getLabel()
             )->withHeader("Location", $location);
     }
